# Add benchmarks for serialization, and make some improvements