Pick some (but not all) low hanging fruit. For example, the encoding
in `github.com/cockroachdb/cockroach/util/encoding` has primitives
better suited to providing an efficient wire encoding.

```
name                                 old time/op    new time/op    delta
Inject_SplitBinary_Empty-8              792ns ± 1%     825ns ± 4%   +4.26%        (p=0.006 n=10+10)
Inject_SplitBinary_100Attributes-8     98.6µs ± 2%    38.0µs ± 2%  -61.48%        (p=0.000 n=10+10)
Extract_SplitBinary_Empty-8             882ns ± 1%     885ns ± 1%     ~            (p=0.287 n=9+10)
Extract_SplitBinary_100Attributes-8     155µs ± 1%      67µs ± 4%  -56.87%        (p=0.000 n=10+10)

name                                 old alloc/op   new alloc/op   delta
Inject_SplitBinary_Empty-8               352B ± 0%      352B ± 0%     ~     (all samples are equal)
Inject_SplitBinary_100Attributes-8     39.9kB ± 0%     6.8kB ± 0%  -83.09%        (p=0.000 n=10+10)
Extract_SplitBinary_Empty-8              224B ± 0%      224B ± 0%     ~     (all samples are equal)
Extract_SplitBinary_100Attributes-8    37.7kB ± 0%    22.5kB ± 0%  -40.38%        (p=0.000 n=10+10)

name                                 old allocs/op  new allocs/op  delta
Inject_SplitBinary_Empty-8               10.0 ± 0%      10.0 ± 0%     ~     (all samples are equal)
Inject_SplitBinary_100Attributes-8      1.22k ± 0%     0.41k ± 0%  -66.28%        (p=0.000 n=10+10)
Extract_SplitBinary_Empty-8              10.0 ± 0%      10.0 ± 0%     ~     (all samples are equal)
Extract_SplitBinary_100Attributes-8     1.21k ± 0%     0.62k ± 0%  -49.09%        (p=0.000 n=10+10)
```