@tschottdorf 2016-02-15
Rebased.

```
name                                           old time/op    new time/op    delta
Span_Empty-8                                      283ns ± 6%     276ns ± 6%     ~            (p=0.217 n=3+10)
Span_100Events-8                                 18.9µs ± 2%    18.4µs ± 1%     ~             (p=0.085 n=3+8)
Span_1000Events-8                                20.5µs ± 6%    18.3µs ± 1%  -10.91%          (p=0.009 n=3+9)
Span_100Tags-8                                   26.0µs ± 3%    26.0µs ± 4%     ~            (p=0.776 n=3+10)
Span_1000Tags-8                                  25.8µs ± 0%    26.0µs ± 5%     ~            (p=0.811 n=3+10)
Span_100Attributes-8                              120µs ± 2%     119µs ± 5%     ~            (p=0.573 n=3+10)
TrimmedSpan_100Events_100Tags_100Attributes-8     125µs ± 7%     122µs ± 5%     ~            (p=0.692 n=3+10)
Inject_SplitText_Empty-8                          590ns ± 4%     328ns ± 6%  -44.51%         (p=0.007 n=3+10)
Inject_SplitText_100Attributes-8                 10.9µs ± 5%     7.7µs ± 5%  -29.41%         (p=0.007 n=3+10)
Inject_SplitBinary_Empty-8                        835ns ± 3%     839ns ± 1%     ~            (p=0.434 n=3+10)
Inject_SplitBinary_100Attributes-8                103µs ± 1%      40µs ± 5%  -61.18%         (p=0.007 n=3+10)
Extract_SplitText_Empty-8                         736ns ± 1%     697ns ± 7%     ~            (p=0.077 n=3+10)
Extract_SplitText_100Attributes-8                 750ns ± 2%     712ns ± 7%     ~            (p=0.084 n=3+10)
Extract_SplitBinary_Empty-8                       852ns ± 2%     876ns ± 2%     ~            (p=0.077 n=3+10)
Extract_SplitBinary_100Attributes-8               160µs ± 1%      71µs ± 2%  -55.59%         (p=0.007 n=3+10)

name                                           old alloc/op   new alloc/op   delta
Span_Empty-8                                     0.00B ±NaN%    0.00B ±NaN%     ~     (all samples are equal)
Span_100Events-8                                 16.1kB ± 0%    16.1kB ± 0%     ~     (all samples are equal)
Span_1000Events-8                                16.1kB ± 0%    16.1kB ± 0%     ~     (all samples are equal)
Span_100Tags-8                                   10.5kB ± 0%    10.5kB ± 0%     ~            (p=0.455 n=3+10)
Span_1000Tags-8                                  10.5kB ± 0%    10.5kB ± 0%     ~            (p=1.021 n=3+10)
Span_100Attributes-8                             10.5kB ± 0%    10.5kB ± 0%   -0.06%         (p=0.042 n=3+10)
TrimmedSpan_100Events_100Tags_100Attributes-8    0.00B ±NaN%    0.00B ±NaN%     ~     (all samples are equal)
Inject_SplitText_Empty-8                           400B ± 0%       64B ± 0%  -84.00%         (p=0.000 n=3+10)
Inject_SplitText_100Attributes-8                 5.85kB ± 0%    0.06kB ± 0%  -98.91%         (p=0.000 n=3+10)
Inject_SplitBinary_Empty-8                         352B ± 0%      352B ± 0%     ~     (all samples are equal)
Inject_SplitBinary_100Attributes-8               39.9kB ± 0%     6.8kB ± 0%  -83.09%         (p=0.000 n=3+10)
Extract_SplitText_Empty-8                        0.00B ±NaN%    0.00B ±NaN%     ~     (all samples are equal)
Extract_SplitText_100Attributes-8                0.00B ±NaN%    0.00B ±NaN%     ~     (all samples are equal)
Extract_SplitBinary_Empty-8                        224B ± 0%      224B ± 0%     ~     (all samples are equal)
Extract_SplitBinary_100Attributes-8              37.7kB ± 0%    22.4kB ± 0%  -40.47%         (p=0.007 n=3+10)

name                                           old allocs/op  new allocs/op  delta
Span_Empty-8                                      0.00 ±NaN%     0.00 ±NaN%     ~     (all samples are equal)
Span_100Events-8                                   8.00 ± 0%      8.00 ± 0%     ~     (all samples are equal)
Span_1000Events-8                                  8.00 ± 0%      8.00 ± 0%     ~     (all samples are equal)
Span_100Tags-8                                     11.0 ± 0%      11.0 ± 0%     ~     (all samples are equal)
Span_1000Tags-8                                    11.0 ± 0%      11.0 ± 0%     ~     (all samples are equal)
Span_100Attributes-8                               11.0 ± 0%      11.0 ± 0%     ~     (all samples are equal)
TrimmedSpan_100Events_100Tags_100Attributes-8     0.00 ±NaN%     0.00 ±NaN%     ~     (all samples are equal)
Inject_SplitText_Empty-8                           4.00 ± 0%      2.00 ± 0%  -50.00%         (p=0.000 n=3+10)
Inject_SplitText_100Attributes-8                   8.00 ± 0%      2.00 ± 0%  -75.00%         (p=0.000 n=3+10)
Inject_SplitBinary_Empty-8                         10.0 ± 0%      10.0 ± 0%     ~     (all samples are equal)
Inject_SplitBinary_100Attributes-8                1.22k ± 0%     0.41k ± 0%  -66.28%         (p=0.000 n=3+10)
Extract_SplitText_Empty-8                         0.00 ±NaN%     0.00 ±NaN%     ~     (all samples are equal)
Extract_SplitText_100Attributes-8                 0.00 ±NaN%     0.00 ±NaN%     ~     (all samples are equal)
Extract_SplitBinary_Empty-8                        10.0 ± 0%      10.0 ± 0%     ~     (all samples are equal)
Extract_SplitBinary_100Attributes-8               1.21k ± 0%     0.62k ± 0%  -49.09%         (p=0.000 n=3+10)
```